Consolidate tiled visualizer tests into per-backend integration files
Each backend (PhysX, Newton) previously required two separate Isaac Sim
app launches: one for the non-tiled integration test and one for the
tiled camera integration test. The four files are now two, with each
file running both test functions under a single shared app launch.

Removes test_visualizer_tiled_integration_physx.py and
test_visualizer_tiled_integration_newton.py.
